@@ -1,62 +1,49 @@
 using System;
-using System.Configuration;
-using System.Collections.Generic;
 using System.IO;
-using System.Linq;
-using System.Threading.Tasks;
-using System.Text;
 using AutoMapper;
-
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.DependencyInjection.Extensions;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.Logging;
-using Microsoft.Extensions.Options;
-using Microsoft.EntityFrameworkCore;
+using FluentValidation;
 using FluentValidation.AspNetCore;
+using IMOMaritimeSingleWindow.Auth;
+using IMOMaritimeSingleWindow.Data;
+using IMOMaritimeSingleWindow.Identity;
+using IMOMaritimeSingleWindow.Identity.Models;
+using IMOMaritimeSingleWindow.ViewModels.Mappings;
+using IMOMaritimeSingleWindow.ViewModels.Validations;
+using log4net;
+using log4net.Config;
+using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Diagnostics;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Http;
-
-using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.AspNetCore.Identity;
-using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.DependencyInjection.Extensions;
+using Microsoft.Extensions.Logging;
+using Microsoft.Extensions.Options;
 using Microsoft.IdentityModel.Tokens;
-
-// Local namespaces
-using IMOMaritimeSingleWindow.Auth;
-using IMOMaritimeSingleWindow.Data;
+using Newtonsoft.Json.Serialization;
+using System.Reflection;
 using IMOMaritimeSingleWindow.Extensions;
 using IMOMaritimeSingleWindow.Helpers;
-using IMOMaritimeSingleWindow.Models;
-using IMOMaritimeSingleWindow.ViewModels.Mappings;
-using IMOMaritimeSingleWindow.Identity;
-using IMOMaritimeSingleWindow.Identity.Models;
 using IMOMaritimeSingleWindow.Repositories;
-using IMOMaritimeSingleWindow.Identity.Stores;
-using Microsoft.AspNetCore.Mvc.Authorization;
-using Microsoft.AspNetCore.Mvc;
-using System.Net.Http.Headers;
-using IMOMaritimeSingleWindow.Identity.Helpers;
-using log4net;
-using log4net.Config;
-using System.Reflection;
 using IMOMaritimeSingleWindow.Filters;
+using IMOMaritimeSingleWindow.Identity.Helpers;
+using IMOMaritimeSingleWindow.Identity.Stores;
 
 namespace IMOMaritimeSingleWindow
 {
     public class Startup
     {
-        static readonly ILog Log = LogManager.GetLogger(System.Reflection.MethodBase.GetCurrentMethod().DeclaringType);
-
+        static readonly ILog Log = LogManager.GetLogger(MethodBase.GetCurrentMethod().DeclaringType);
 
         public Startup(IHostingEnvironment env)
         {
             var logRepository = LogManager.GetRepository(Assembly.GetEntryAssembly());
             XmlConfigurator.Configure(logRepository, new FileInfo("log4net.config"));
-
-
 
             Configuration = new ConfigurationBuilder()
                 .SetBasePath(env.ContentRootPath)
@@ -66,257 +53,135 @@ namespace IMOMaritimeSingleWindow
                 .AddEnvironmentVariables()
                 .Build();
             HostingEnvironment = env;
-
         }
-
 
         public IConfigurationRoot Configuration { get; }
         public IHostingEnvironment HostingEnvironment { get; }
-        public AuthorizationPolicy AuthorizationPolicy { get; internal set; }
+public void ConfigureServices(IServiceCollection services)
+{
+    services.AddMvc(opts =>
+    {
+        opts.Filters.Add(new Log4NetExceptionFilter());
+        opts.EnableEndpointRouting = false; // Disable Endpoint Routing
+    });
 
-        // This method gets called by the runtime. Use this method to add services to the container.
-        // For more information on how to configure your application, visit https://go.microsoft.com/fwlink/?LinkID=398940
-        public void ConfigureServices(IServiceCollection services)
-
+    // Add FluentValidation and Newtonsoft.Json
+    services.AddControllers()
+        .AddNewtonsoftJson(options =>
         {
-            // Wrap the entire configuration routine in a try/catch since the logging filter won't catch any errors that occur here
-            try
-            {
-
-                // Add exception logging
-                services.AddMvc(opts =>
-                    opts.Filters.Add(new Log4NetExceptionFilter())
-                );
-
-                services.TryAddTransient<IHttpContextAccessor, HttpContextAccessor>();
-                //Configure CORS with different policies
-                //services.AddCors(options =>
-                //{
-                //    options.AddPolicy("AllowLocalhost",
-                //        b => {
-                //            b.WithOrigins(new string[] {
-                //                "http://localhost:4200",
-                //                "https://localhost:4200"
-                //            });
-                //            b.WithMethods(new string[]
-                //            {
-                //                "GET", "OPTIONS", "POST", "UPDATE"
-                //            });
-                //            b.AllowAnyHeader();
-                //        });
-
-                //    //Brute force policy if all else fails
-                //    //NB: Only ever use in development!
-                //    options.AddPolicy("AllowAllAny",
-                //        b => b.AllowAnyOrigin().AllowAnyHeader().AllowAnyMethod());
-                //});
-
-                //Configure database context
-                string connectionStringOpenSSN = "";
-                if (HostingEnvironment.IsStaging())
-                {
-                    // Configure in-memory database
-                }
-                else
-                {
-                    // Use real database
-                    connectionStringOpenSSN = Configuration.GetConnectionString("OpenSSN");
-                }
-
-                connectionStringOpenSSN = Configuration.GetConnectionString("OpenSSN");
-                var dbOptions = new DbContextOptionsBuilder<open_ssnContext>().UseNpgsql(connectionStringOpenSSN).Options;
-                services.AddEntityFrameworkNpgsql().AddDbContext<open_ssnContext>(options => options.UseNpgsql(connectionStringOpenSSN));
-
-                // Configure email service
-                services.ConfigureEmailSenderOptions(Configuration);
-                services.ConfigureSendGridOptions(Configuration);
-
-                services.AddEmailSender();
-
-
-                //Configure identity services
-                var builder = services.AddIdentity<ApplicationUser, ApplicationRole>(options =>
-                {
-                // configure identity options
-                options.Password.RequireDigit = false;
-                    options.Password.RequireLowercase = false;
-                    options.Password.RequireUppercase = false;
-                    options.Password.RequireNonAlphanumeric = false;
-                    options.Password.RequiredLength = 4;
-
-                // Lockout options
-                options.Lockout.DefaultLockoutTimeSpan = TimeSpan.FromMinutes(30);
-                    options.Lockout.MaxFailedAccessAttempts = 10;
-                    options.Lockout.AllowedForNewUsers = true;
-
-                //Email options
-                options.SignIn.RequireConfirmedEmail = true;
-                });
-
-                builder.AddDefaultTokenProviders();
-
-                //builder.AddSignInManager<SignInManager<ApplicationUser>>()
-                //builder.AddUserManager<ApplicationUserManager>()
-                //.AddRoleManager<ApplicationRoleManager>();
-
-
-                var serviceProvider = services.BuildServiceProvider();
-                var context = serviceProvider.GetService<open_ssnContext>();
-                if (context == null) throw new Exception("no service for open_ssnContext found!");
-
-                //services.AddSingleton<IUnitOfWork<Guid>, UnitOfWork>();
-                // services.AddAutoMapper();
-                //var config = new MapperConfiguration(cfg => cfg.AddProfiles(typeof(Startup)));
-
-                //services.TryAddScoped<UserManager>();
-                //services.TryAddScoped<ApplicationRoleManager>();
-
-                // Tip from https://stackoverflow.com/a/42298278
-                var config = new MapperConfiguration(cfg =>
-               {
-                   cfg.AddProfile<IdentityEntitiesToModelsMappingProfile>();
-                   cfg.AddProfile<ViewModelToEntityMappingProfile>();
-               });
-                services.AddSingleton<IMapper>(s => config.CreateMapper());
-                services.TryAddScoped<IUnitOfWork<Guid>>(ctx => new UnitOfWork(new open_ssnContext(dbOptions)));
-                serviceProvider = services.BuildServiceProvider();
-
-                var automapper = serviceProvider.GetService<IMapper>();
-
-                services.AddSingleton<IUserStoreHelper>(ctx => new UserStoreHelper(automapper));
-                services.TryAddScoped<IdentityErrorDescriber>();
-                serviceProvider = services.BuildServiceProvider();
-
-                var userStoreHelper = serviceProvider.GetService<IUserStoreHelper>();
-                var unitofwork = (UnitOfWork)serviceProvider.GetService<IUnitOfWork<Guid>>();
-                var identityErrorDescriber = serviceProvider.GetService<IdentityErrorDescriber>();
-
-                builder.AddUserManager<UserManager>()
-                .AddRoleManager<ApplicationRoleManager>();
-
-
-                services.TryAddScoped<IUserStore<ApplicationUser>>(ctx =>
-                    new UserStore
-                    (
-                        identityErrorDescriber,
-                        new UnitOfWork(new open_ssnContext(dbOptions)),
-                        new RoleStore(new UnitOfWork(new open_ssnContext(dbOptions)), automapper),
-                        userStoreHelper,
-                        automapper
-                    )
-                );
-
-                services.TryAddScoped<IRoleStore<ApplicationRole>>(ctx => new RoleStore(
-                    new UnitOfWork(new open_ssnContext(dbOptions)),
-                    automapper)
-                );
-
-                services.TryAddScoped<IDbContext>(ctx => new open_ssnContext(dbOptions));
-
-                //services.TryAddScoped<IRoleStore<ApplicationRole>>(ctx => new RoleStore(unitofwork, automapper));
-                //var roleStore = serviceProvider.GetService<RoleStore>();
-                //services.TryAddScoped<IUserStore<ApplicationUser>>(ctx => new UserStore(unitofwork, roleStore, automapper));
-
-                serviceProvider = services.BuildServiceProvider();
-
-                var myUserManager = serviceProvider.GetService<UserManager<ApplicationUser>>();
-                //var myRoleManager = serviceProvider.GetService<RoleManager<ApplicationRole>>();
-
-                // Additional manager separate from ASP NET Identity
-                //services.TryAddScoped(ctx => new UserRoleManager<ApplicationUser, Guid, ApplicationRole, Guid>(myUserManager, myRoleManager));
-
-
-                //Overriding service
-                services.Replace(ServiceDescriptor.Scoped<IUserValidator<ApplicationUser>, CustomUserValidator<ApplicationUser>>());
-                //services.Replace(ServiceDescriptor.Scoped<IPasswordHasher<ApplicationUser>, CustomPasswordHasher>());
-                // Additional manager separate from ASP NET Identity
-                //services.TryAddScoped(ctx => new UserRoleManager<ApplicationUser, Guid, ApplicationRole, Guid>(myUserManager, myRoleManager));
-
-
-                // Custom services
-
-                //services.AddScoped<IUserClaimsPrincipalFactory<ApplicationUser>, ApplicationClaimsPrincipalFactory>();
-
-                services.AddSingleton<IJwtFactory, JwtFactory>();
-
-
-                //See IMOMaritimeSingleWindow.Extensions.IServiceCollections.cs for implementation
-                services.AddJWTOptions(Configuration);
-                services.AddAuthorizationPolicies();
-
-                if (HostingEnvironment.IsProduction())
-                {
-                    services.AddMvc(opts =>
-                        opts.Filters.Add(new RequireHttpsAttribute())
-                    );
-                }
-
-
-
-                services.AddMvc()
-                    .AddFluentValidation(fv => fv.RegisterValidatorsFromAssemblyContaining<Startup>());
-
-
-
-
-
-                /*services.AddSingleton<IEmailSender, EmailSender>();
-                services.Configure<AuthMessageSenderOptions>(Configuration);*/
-
-                /*var b = services.AddMvc(
-                    options => {
-                        var defaultPolicy = new AuthorizationPolicyBuilder(new[] { JwtBearerDefaults.AuthenticationScheme, IdentityConstants.ApplicationScheme })
-                        .RequireAuthenticatedUser().Build();
-                        options.Filters.Add(new AuthorizeFilter(defaultPolicy));
-
-                    });
-                */
-
-                // Fix for json self-referencing loop bug:
-                services.AddMvc().AddJsonOptions(
-              options => options.SerializerSettings.ReferenceLoopHandling =
-              Newtonsoft.Json.ReferenceLoopHandling.Ignore
-            );
-            }
-            catch(Exception e)
-            {
-                Log.Error(e);
-                if (e.InnerException != null)
-                {
-                    Log.Error(e.InnerException);
-                }
-            }
-        }
-
-        // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+            options.SerializerSettings.ContractResolver = new DefaultContractResolver();
+            options.SerializerSettings.ReferenceLoopHandling = Newtonsoft.Json.ReferenceLoopHandling.Ignore;
+        })
+        .AddFluentValidation(fv =>
         {
-            app.Use(async (context, next) =>
-            {
-                await next();
-                if (context.Response.StatusCode == 404 &&
-                   !Path.HasExtension(context.Request.Path.Value) &&
-                   !context.Request.Path.Value.StartsWith("/api/", StringComparison.Ordinal))
-                {
-                    context.Request.Path = "/index.html";
-                    await next();
-                }
-            });
+            fv.RegisterValidatorsFromAssembly(Assembly.GetExecutingAssembly());
+        });
 
-            if (!env.IsProduction())
-            {
-                app.UseCors(builder => builder.AllowAnyOrigin().AllowAnyMethod().AllowAnyHeader());
-            }
+    services.TryAddTransient<IHttpContextAccessor, HttpContextAccessor>();
 
-            //app.UseCors("AllowAllAny");
-            // app.UseCors(policyName: "AllowLocalhost");
+    // Configure database context
+    string connectionStringOpenSSN = Configuration.GetConnectionString("OpenSSN");
+    services.AddDbContext<open_ssnContext>(options =>
+        options.UseNpgsql(connectionStringOpenSSN));
 
-            // IMPORTANT! UseAuthentication() must be called before UseMvc()
+    // Register the DbContext as IDbContext
+    services.AddScoped<IDbContext>(provider => provider.GetService<open_ssnContext>());
 
-            app.UseAuthentication();
-            app.UseMvcWithDefaultRoute();
-            app.UseDefaultFiles();
-            app.UseStaticFiles();
-        }
+    // Register UnitOfWork
+    services.AddScoped<IUnitOfWork<Guid>, UnitOfWork>();
+
+    // Configure identity services
+    services.AddIdentity<ApplicationUser, ApplicationRole>(options =>
+    {
+        // configure identity options
+        options.Password.RequireDigit = false;
+        options.Password.RequireLowercase = false;
+        options.Password.RequireUppercase = false;
+        options.Password.RequireNonAlphanumeric = false;
+        options.Password.RequiredLength = 4;
+
+        // Lockout options
+        options.Lockout.DefaultLockoutTimeSpan = TimeSpan.FromMinutes(30);
+        options.Lockout.MaxFailedAccessAttempts = 10;
+        options.Lockout.AllowedForNewUsers = true;
+
+        // Email options
+        options.SignIn.RequireConfirmedEmail = true;
+    })
+    .AddEntityFrameworkStores<open_ssnContext>()
+    .AddDefaultTokenProviders();
+
+    // AutoMapper configuration
+    var config = new MapperConfiguration(cfg =>
+    {
+        cfg.AddProfile<IdentityEntitiesToModelsMappingProfile>();
+        cfg.AddProfile<ViewModelToEntityMappingProfile>();
+    });
+    services.AddSingleton<IMapper>(s => config.CreateMapper());
+    services.TryAddScoped<IUnitOfWork<Guid>>(ctx => new UnitOfWork(ctx.GetRequiredService<open_ssnContext>()));
+
+    services.AddSingleton<IUserStoreHelper, UserStoreHelper>(); 
+    services.TryAddScoped<IdentityErrorDescriber>();
+
+    services.AddScoped<IUserStore<ApplicationUser>, UserStore>(); 
+    services.AddScoped<IRoleStore<ApplicationRole>, RoleStore>(); 
+
+    // Overriding service
+    services.Replace(ServiceDescriptor.Scoped<IUserValidator<ApplicationUser>, CustomUserValidator<ApplicationUser>>());
+
+    // Custom services
+    services.AddSingleton<IJwtFactory, JwtFactory>();
+
+    // See IMOMaritimeSingleWindow.Extensions.IServiceCollections.cs for implementation
+    services.AddJWTOptions(Configuration);
+    services.AddAuthorizationPolicies();
+
+    if (HostingEnvironment.IsProduction())
+    {
+        services.AddMvc(opts => opts.Filters.Add(new RequireHttpsAttribute())); // Add using Microsoft.AspNetCore.Mvc;
     }
 }
+public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+{
+    app.Use(async (context, next) =>
+    {
+        var userManager = context.RequestServices.GetService<UserManager<ApplicationUser>>();
+        var signInManager = context.RequestServices.GetService<SignInManager<ApplicationUser>>();
+        var roleManager = context.RequestServices.GetService<RoleManager<ApplicationRole>>();
+
+        if (userManager == null)
+        {
+            Console.WriteLine("UserManager<ApplicationUser> is not registered.");
+        }
+        else
+        {
+            Console.WriteLine("UserManager<ApplicationUser> is registered.");
+        }
+
+        if (signInManager == null)
+        {
+            Console.WriteLine("SignInManager<ApplicationUser> is not registered.");
+        }
+        else
+        {
+            Console.WriteLine("SignInManager<ApplicationUser> is registered.");
+        }
+
+        if (roleManager == null)
+        {
+            Console.WriteLine("RoleManager<ApplicationRole> is not registered.");
+        }
+        else
+        {
+            Console.WriteLine("RoleManager<ApplicationRole> is registered.");
+        }
+
+        await next.Invoke();
+    });
+
+    app.UseAuthentication();
+    app.UseMvcWithDefaultRoute();
+    app.UseDefaultFiles();
+    app.UseStaticFiles();
+}
+    }}
