@@ -19,8 +19,8 @@ namespace IMOMaritimeSingleWindow.ViewModels.Validations
     {
         public CredentialsViewModelValidator()
         {
-            RuleFor(vm => vm.UserName).NotEmpty().WithMessage("Username cannot be empty");
-            RuleFor(vm => vm.Password).NotEmpty().WithMessage("Password cannot be empty");
+            RuleFor(x => x.UserName).NotEmpty().WithMessage("User name cannot be empty");
+            RuleFor(x => x.Password).NotEmpty().WithMessage("Password cannot be empty");
         }
     }
 }
